+++ src/main/resources/schema.sql
@@ -4,7 +4,7 @@ drop table if exists user;
 
 create table if not exists role
 (
-    role_id bigint auto_increment primary key,
+    role_id int auto_increment primary key,
     role    varchar(255) not null unique
 );
 
@@ -21,7 +21,7 @@ create table if not exists user
 create table if not exists user_role
 (
     user_id bigint not null,
-    role_id bigint not null,
+    role_id int not null,
     primary key (user_id, role_id),
     foreign key (user_id) references user (user_id),
     foreign key (role_id) references role (role_id)
